Create a common base class for HConfig and HConfigChild
